Share day stats merge helper
