+++ src/core/mod.rs
@@ -8,7 +8,7 @@ mod types;
 
 pub(crate) use aggregator::{
     aggregate_blocks, aggregate_daily, aggregate_projects, aggregate_sessions,
-    aggregate_sessions_map, format_project_name,
+    aggregate_sessions_map, format_project_name, merge_day_stats,
 };
 pub(crate) use dedup::DedupAccumulator;
 pub(crate) use tool_aggregator::aggregate_tools;
